Extract BasicFormField interface and remove isDirty property
This change prepares the foundation for TextFieldState integration by separating value-agnostic operations from the
existing structure. Additionally, the isDirty property has been removed due to its incompatibility with TextFieldState's
input mechanism, which will streamline the integration process and ensure consistent behavior across the form field
implementations.
